+++ server/federation/index.js
@@ -36,12 +36,29 @@ router.get('/m/:event_id:json(.json)?', async (req, res) => {
   if (!event) { return res.status(404).send('Not found') }
   const eventAp = event.toAP(settingsController.settings)
   eventAp['@context'] = [
-    "https://www.w3.org/ns/activitystreams",
+    'https://www.w3.org/ns/activitystreams',
+    'https://w3id.org/security/v1',
     {
-      "sc": "http://schema.org/",
-      "address": {
-        "@id": "sc:address",
-        "@type": "sc:Text"
+      toot: 'http://joinmastodon.org/ns#',
+
+      // A property-value pair, e.g. representing a feature of a product or place. We use this to publish this very same instance
+      // https://docs.joinmastodon.org/spec/activitypub/#PropertyValue
+      schema: 'http://schema.org#',
+      ProperyValue: 'schema:PropertyValue',
+      value: 'schema:value',
+
+      // https://docs.joinmastodon.org/spec/activitypub/#discoverable
+      "discoverable": "toot:discoverable",
+
+      // https://docs.joinmastodon.org/spec/activitypub/#Hashtag
+      "Hashtag": "https://www.w3.org/ns/activitystreams#Hashtag",
+
+      manuallyApprovesFollowers: 'as:manuallyApprovesFollowers',
+
+      // focal point - https://docs.joinmastodon.org/spec/activitypub/#focalPoint
+      "focalPoint": {
+        "@container": "@list",
+        "@id": "toot:focalPoint"
       }
     }
   ]