+++ server/helpers.js
@@ -279,6 +279,7 @@ module.exports = {
         log.debug('[FEDI] APRedirect for %s', event.title)
         return res.redirect(`/federation/m/${event.id}`)
       }
+      log.warn('[FEDI] Accept JSON but event not found: %s', req.params.slug)
     }
     next()
   },