+++ vuetify.options.js
@@ -10,7 +10,6 @@ export default ({ res, nuxtState }) => {
 
   return {
     lang: { locales: { ca, cs, de, en, es, eu, fr, gl, it, nb, nl, pl, pt, sk, tr, ro, ru, zhHans } },
-    treeShake: true,
     theme: {
       options: {
         customProperties: false,
@@ -23,6 +22,5 @@ export default ({ res, nuxtState }) => {
         light: settings.light_colors
       }
     },
-    defaultAssets: false
   }
 }